google-cloud/b24-spa-1106-sync/b24-spa-1106-http/requirements.txt
functions-framework==3.*
flask>=2.0.0
google-cloud-pubsub>=2.18.0
